+++ cmake/packaging.cmake
@@ -20,6 +20,10 @@ install(FILES "${CMAKE_SOURCE_DIR}/dependencies/CLIUtils/LICENSE"
     DESTINATION "${CMAKE_INSTALL_DATADIR}/doc/${CMAKE_PROJECT_NAME}"
     RENAME "CLI11-LICENSE.txt"
 )
+install(FILES "${CMAKE_SOURCE_DIR}/dependencies/elfio/LICENSE.txt"
+    DESTINATION "${CMAKE_INSTALL_DATADIR}/doc/${CMAKE_PROJECT_NAME}"
+    RENAME "ELFIO-LICENSE.txt"
+)
 
 # `file(ARCHIVE_CREATE COMPRESSION Gzip)` creates files with
 # timestamps without a way of turning them off.  Instead, just use